Simplify customer filtering:
Filter the customers-array when rendering, not when updating the state.
This removes the need for `componentWillReceiveProps` and puts all the
filtering+sorting logic for customers in one place. `state` is now only
used for storing the view state (the filter and sorting value) and has
no data at all.

The disadvantage is that we are constantly re-filtering the customers,
but that will usually not be a performance issue. And if it does become
a performance issue we can solve it in a different way.