OrdersTable: Move row rendering inline
I understand the need for having a separate OrderRow, but personally I
always start with having an inline renderer for rows, and then refactor
it out into its own file when it grows.

Many will probably disagree with me, but I think the pattern of having
multiple render-methods inside a single component is useful to break up
a big render() without having to define separate components. One of the
advantages of React's virtual DOM is that it is just regular JavaScript
values and you can structure it using functions.